backend/requirements.txt
Django==5.1.7
MarkupSafe==2.1.5
PyJWT==2.8.0
PyYAML==6.0.1
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cryptography==44.0.1
defusedxml==0.7.1
docker==7.1.0
django-allauth==65.2.0
django-cors-headers==4.3.1
djangorestframework==3.15.2
clickhouse-connect==0.8.17
flyql @ git+https://github.com/iamtelescope/flyql.git@054ab49e885d3abe15967e714367ce690b872166#subdirectory=python
gunicorn==23.0.0
idna==3.7
jsonschema==4.23.0
kubernetes==31.0.0
oauthlib==3.2.2
pycparser==2.22
python3-openid==3.2.0
pytz==2024.1
python-dateutil==2.9.0.post0
requests-oauthlib==2.0.0
requests==2.32.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.5.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
cachetools